refactor(tokenomics): remove dead attribution UI + legacy channels/types/preload; final cleanup
