+++ src/shared/types.ts
@@ -345,75 +345,6 @@ export interface TokenomicsSessionRecord {
   configLabel?: string
 }
 
-export interface TokenomicsDailyAggregate {
-  date: string
-  totalCostUsd: number
-  totalTokens: number
-  messageCount: number
-  sessionCount: number
-  totalDurationMs: number
-  avgCostPerHour: number
-  byModel: Record<string, { costUsd: number; inputTokens: number; outputTokens: number }>
-  /** Per-account daily rollup, keyed by canonical accountEmail (unattributed sessions
-   *  fall under '__unattributed__' so the axis reconciles to the day total). Optional
-   *  for back-compat: rebuilt on load, so persisted pre-account aggregates lack it. */
-  byAccount?: Record<string, { costUsd: number; inputTokens: number; outputTokens: number }>
-}
-
-export interface TokenomicsData {
-  sessions: Record<string, TokenomicsSessionRecord>
-  dailyAggregates: Record<string, TokenomicsDailyAggregate>
-  lastSyncTimestamp: number
-  totalCostUsd: number
-  seedComplete: boolean
-  // Extra spend tracking (from Anthropic API via statusline)
-  extraSpend?: {
-    enabled: boolean
-    usedUsd: number
-    limitUsd: number
-    lastUpdated: number // epoch ms
-  }
-  // Rate limit tracking (from Anthropic API via statusline)
-  rateLimits?: {
-    fiveHour?: number    // utilization percentage
-    sevenDay?: number    // utilization percentage
-    lastUpdated: number
-  }
-  // P6: Codex review (Claude-driven) -- per-day aggregates from
-  // <resourcesDir>/tokenomics/codex-review-by-day.json. Distinct from
-  // interactive Codex usage already in dailyAggregates.
-  codexReviewByDay?: Record<string, {
-    reviewCount: number
-    totalInputTokens: number
-    totalOutputTokens: number
-  }>
-}
-
-export interface TokenomicsSyncProgress {
-  phase: 'scanning' | 'processing' | 'complete'
-  totalFiles: number
-  processedFiles: number
-  currentFile?: string
-}
-
-// -- P8: Attribution wizard --
-
-export type AttributionPayload = {
-  sessionIds: string[]
-  assignment:
-    | { type: 'email'; email: string }
-    | { type: 'mixed' }
-    | { type: 'clear' }
-}
-
-export interface UnattributedSessionGroup {
-  groupId: string                     // configId or '__no-config__'
-  groupLabel: string                  // e.g. "This App Dev" or "(no config)"
-  sessionIds: string[]
-  totalCostUsd: number
-  suggestedEmail: string | null       // null when before earliest backup
-}
-
 // -- Codex Review (P6) --
 
 export interface CodexReviewRateLimitWindow {
